requirements.txt
pandas>=2.0
numpy>=1.24
yfinance>=0.2
requests>=2.28
beautifulsoup4>=4.12
openpyxl>=3.1
plotly>=5.15
streamlit>=1.28
pyyaml>=6.0
scipy>=1.10
statsmodels>=0.14
